refactor: CalculatorTest 코드를 main/Calculator/Calculator.java로 분리
